requirements.txt
openai==1.101.0
anthropic
pageindex
requests
pymupdf==1.26.4
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.11.0
pyyaml==6.0.2
pydantic>=2.0
hypothesis
pytest
pytest-asyncio
